Learning to add agent instructions
