+++ PandaBattleship.API.Tests/OpenApiEndpointTests.cs
@@ -1,0 +1,27 @@
+﻿using Microsoft.AspNetCore.Mvc.Testing;
+using System.Net;
+
+namespace PandaBattleship.API.Tests;
+
+public class OpenApiEndpointTests : IClassFixture<WebApplicationFactory<Program>>
+{
+    private readonly WebApplicationFactory<Program> _factory;
+
+    public OpenApiEndpointTests(WebApplicationFactory<Program> factory)
+    {
+        _factory = factory;
+    }
+
+    [Fact]
+    public async Task GetOpenApiDocument_ReturnsSuccess()
+    {
+        var client = _factory.CreateClient();
+
+        var response = await client.GetAsync("/openapi/v1.json");
+
+        response.EnsureSuccessStatusCode();
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        var contentType = response.Content.Headers.ContentType?.MediaType;
+        Assert.Equal("application/json", contentType);
+    }
+}
